+++ R/contrast.R
@@ -342,8 +342,10 @@ contrast.emmGrid = function(object, method = "eff", interaction = FALSE,
     }
     else if (is.character(method)) {
         fn = paste(method, "emmc", sep=".")
-        method = if (exists(fn, mode="function", envir = parent.frame())) 
-            get(fn, envir = parent.frame()) 
+        if (exists(fn, mode = "function")) 
+            method = get(fn) 
+        else if (exists(fn, mode = "function", envir = parent.frame()))
+            method = get(fn, envir = parent.frame()) 
         else 
             stop(paste("Contrast function '", fn, "' not found", sep=""))
         if(!missing(wts) && any(is.na(wts))) {