+++ R/emm-list.R
@@ -59,7 +59,7 @@
 #' @order 1
 #' @examples
 #' mod <- lm(conc ~ source, data = pigs)
-#' obj <- emmeans(pigs.lm, pairwise ~ source)
+#' obj <- emmeans(mod, pairwise ~ source)
 #' 
 #' linfct(obj)
 #'